Desativação de um produto; Validações e ajustes em atualização de produto

+++ EcommerceCandyHill.Domain/Interfaces/Repositories/IBaseRepository.cs
@@ -17,6 +17,6 @@ namespace EcommerceCandyHill.Domain.Interfaces.Repositories
 
         void Update(Product product);
 
-        void Delete(long id);
+        void Delete(int id);
     }
 }
